Created a "Better" gui
Actually it is absolute trash, but I'm fixing it =)

I created a copy so I wouldn't mess everything up since I don't know how to use branches yet.

What I did was change from typing the name of the item to a number, however the order is random so you buy whatever the program wants, good for whoever enjoys a good gambling.