loop shop and keep track of items
Changed so that program will run until you cancel it or run out of money.

Added a tracker, so that I can know how many items I got, and which items  they were.

Made a 'while' loop inside the main function that will keep the function running until you either run out of money or cancel it yourself.

==========

The design is uglier and I don't like that but not much can be done :/